% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractScanInfo.R
\name{extractScanInfo}
\alias{extractScanInfo}
\title{A function that extracts scan header information from a mzML file}
\usage{
extractScanInfo(path)
}
\arguments{
\item{path}{The path to the mzML file.}
}
\value{
output A data frame
}
\description{
This function extracts informations from a mzML file exported from 
ThermoRAW file with msConvert software. The information includes 
scan number, MS level, scanfilter message, total ion current, base peak, 
injection time etc. This function can be used to check injection time with 
the intention of optimizing the acquisition conditions.
}

